+++ NEWS.md
@@ -50,7 +50,7 @@
 * Adjusted `tbl_hierarchical_incidence_rate()` to handle first event as well as
 all events (#217)
 
-* Adjusted `tbl_baseline_chg()` to allow for custiomization of display statistics (#221)
+* Adjusted `tbl_baseline_chg()` to allow for custiomization of display statistics (#222)
 
 
 # crane 0.3.1